Modularize Data Cafe.
