+++ conf/datacafe.properties
@@ -52,3 +52,12 @@ completeMySQLUrl=localhost:3306
 mySQLUserName=root
 mySQLPassword=root
 additionalMySQLConf=?useUnicode=true&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC
+
+
+
+# Hazelcast Configurations
+hazelcastXml=conf/hazelcast.xml
+
+# Name of the execution cluster. Default is "main". Change this for parallel simulations.
+# Uncomment and use different cluster names for multi-tenanted deployments.
+mainCluster=datacafe
