+++ modules/datacafe-client/src/main/java/edu/emory/bmi/datacafe/client/samples/package-info.java
@@ -1,0 +1,4 @@
+/*
+ * Samples for the client bundle.
+ */
+package edu.emory.bmi.datacafe.client.samples;
