+++ modules/datacafe-core/pom.xml
@@ -17,14 +17,6 @@
     <version>1.0-SNAPSHOT</version>
   </parent>
   
-  <artifactId>datacafe-server</artifactId>
-  <name>datacafe-server</name>
-
-  <dependencies>
-      <dependency>
-          <groupId>edu.emory.bmi</groupId>
-          <artifactId>datacafe-core</artifactId>
-      </dependency>
-  </dependencies>
-
+  <artifactId>datacafe-core</artifactId>
+  <name>datacafe-core</name>
 </project>
