+++ modules/datacafe-core/src/main/java/edu/emory/bmi/datacafe/client/conf/DatacafeConstants.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.emory.bmi.datacafe.constants;
+package edu.emory.bmi.datacafe.client.conf;
 
 /**
  * The common constants of Data Cafe
