+++ modules/datacafe-core/src/main/java/edu/emory/bmi/datacafe/client/conf/package-info.java
@@ -1,0 +1,4 @@
+/*
+ * Core configurations package.
+ */
+package edu.emory.bmi.datacafe.client.conf;
