+++ modules/datacafe-core/src/main/java/edu/emory/bmi/datacafe/client/hazelcast/HzConstants.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.emory.bmi.datacafe.hazelcast;
+package edu.emory.bmi.datacafe.client.hazelcast;
 
 /**
  * Constants specific to Hazelcast
