+++ modules/datacafe-core/src/main/java/edu/emory/bmi/datacafe/client/hazelcast/package-info.java
@@ -1,0 +1,4 @@
+/*
+ * Core Hazelcast Integration for the In-Memory Data Grid for Data Cafe.
+ */
+package edu.emory.bmi.datacafe.impl.hazelcast;
