+++ modules/datacafe-samples/src/main/java/edu/emory/bmi/datacafe/impl/hazelcast/package-info.java
@@ -1,0 +1,4 @@
+/*
+ * Samples testing: Hazelcast to store the user queries dynamically
+ */
+package edu.emory.bmi.datacafe.impl.hazelcast;
