+++ modules/datacafe-samples/src/main/java/edu/emory/bmi/datacafe/impl/mongo/clinical/main/Executor.java
@@ -15,8 +15,8 @@
  */
 package edu.emory.bmi.datacafe.impl.mongo.clinical.main;
 
-import edu.emory.bmi.datacafe.core.CoreExecutorEngine;
-import edu.emory.bmi.datacafe.core.DataSourcesRegistry;
+import edu.emory.bmi.datacafe.client.CoreExecutorEngine;
+import edu.emory.bmi.datacafe.client.DataSourcesRegistry;
 import edu.emory.bmi.datacafe.hdfs.HdfsConnector;
 import edu.emory.bmi.datacafe.mongo.MongoConnector;
 import org.bson.Document;
