+++ modules/datacafe-samples/src/main/java/edu/emory/bmi/datacafe/impl/mongo/parallel/main/PhysioNetSeqBenchmark.java
@@ -15,11 +15,10 @@
  */
 package edu.emory.bmi.datacafe.impl.mongo.parallel.main;
 
-import edu.emory.bmi.datacafe.core.CoreExecutorEngine;
-import edu.emory.bmi.datacafe.core.DataSourcesRegistry;
+import edu.emory.bmi.datacafe.client.CoreExecutorEngine;
+import edu.emory.bmi.datacafe.client.DataSourcesRegistry;
 import edu.emory.bmi.datacafe.hdfs.HdfsConnector;
 import edu.emory.bmi.datacafe.impl.mongo.clinical.main.ExecutorRandomID;import edu.emory.bmi.datacafe.mongo.MongoConnector;
-import edu.emory.bmi.datacafe.mongo.MongoHDFSIntegrator;
 import edu.emory.bmi.datacafe.mongo.MongoIntegratedConnector;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
