+++ modules/datacafe-server/src/main/java/edu/emory/bmi/datacafe/client/package-info.java
@@ -3,4 +3,4 @@
  * Core of the Data Cafe Engine..
  */
 
-package edu.emory.bmi.datacafe.core;
+package edu.emory.bmi.datacafe.client;
