+++ modules/datacafe-server/src/main/java/edu/emory/bmi/datacafe/hdfs/HdfsUtil.java
@@ -17,7 +17,7 @@ package edu.emory.bmi.datacafe.hdfs;
 
 import edu.emory.bmi.datacafe.conf.ConfigReader;
 import edu.emory.bmi.datacafe.constants.HDFSConstants;
-import edu.emory.bmi.datacafe.core.CoreExecutorEngine;
+import edu.emory.bmi.datacafe.client.CoreExecutorEngine;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
