+++ modules/datacafe-server/src/main/java/edu/emory/bmi/datacafe/mysql/MySQLConnector.java
@@ -17,7 +17,7 @@ package edu.emory.bmi.datacafe.mysql;
 
 import edu.emory.bmi.datacafe.conf.ConfigReader;
 import edu.emory.bmi.datacafe.constants.SqlConstants;
-import edu.emory.bmi.datacafe.core.AbstractDataSourceConnector;
+import edu.emory.bmi.datacafe.client.AbstractDataSourceConnector;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
